Issue #67. Test multiple audience field with EFQ
Fixes #67.

Please notice that, because of a bug in OG reported in #67, we have to look for an alternative way to test EntityFieldQuery with multiple OG audience fields, basically using two EFQs and intersecting them.

If this issue is fixed, it'd be good to revert to the original way of testing this.